Add Hack build target
This splits the module import, which is used for the Pester target, into
an Import target. Since importing the module is only dependent on the
Init target, the Hack target has been created to add the additional
Build step.

The purpose of adding this is to allow for quick hacking on an
individual function. Since the Test target runs all configured tests, it
can take an unreasonable amount of time for verifying small changes.
After building and importing with this target, Pester can be invoked for
the singular active test file.