Improve perf test setup involving git
These tests set up a remote template repository which is copied into a
new directory for each test. However, the local repository was cloned
for each test separately.

With this change, tests take the same approach for the local repository,
where the clone operation happens into a template repository at the
beginning and the template is copied for each test. The only additional
step is for each test to update the remote URL to point to the new
remote rather than the template remote.