Exclude leaves on multiple build artifact matches
When multiple directories within a path match as a build artifact
directory, only the top-most directory should be deleted. This fixes the
issue where directories are deleted within node package directories, as
we do not want to implicitly force a re-download of that many packages.

For path 1 below, directory A is specifically excluded from deletion, so
we do not want to delete directory B and create issues with the
associated package.

                   A                           B
                   |                           |
1: /c/Projects/Foo/node_modules/sample-package/bin/

For path 2 below, since directory A is already being deleted, we do not
want or need to attempt deletion of directory B.

                   A                       B
                   |                       |
2: /c/Projects/Foo/artifacts/sample-output/bin/