Use end block when writing final host segment
The host segment writing function was previously keeping a count of the
written segments and outputing the newline once the final segment count
had been reached.

This has been simplified to use the end block provided by the
ForEach-Object cmdlet. This way, no count needs to be tracked throughout
the loop.